Stop database before exiting when starting fails
In case the `as_token` or the `/register` request are not accepted, the application hangs indefinitely (when using SQLite).
The reason for this is because [this thread](https://github.com/omnilib/aiosqlite/blob/main/aiosqlite/core.py#L84) in `aiosqlite` will never exit unless the connection is closed.
This can be solved by closing the database before exiting here.